SDK-1547. Add getDownloadUrl for apps to get direct download links

See merge request sdk/sdk!2829

+++ src/commands.cpp
@@ -621,7 +621,7 @@ bool CommandDirectRead::procresult(Result r)
 }
 
 // request temporary source URL for full-file access (p == private node)
-CommandGetFile::CommandGetFile(MegaClient *client, const byte* key, unsigned keySize,
+CommandGetFile::CommandGetFile(MegaClient *client, const byte* key, size_t keySize,
                                handle h, bool p, const char *privateauth,
                                const char *publicauth, const char *chatauth,
                                bool singleUrl, Cb &&completion)
